requirements.txt
numpy
pika==1.3.0
psutil
scikit_learn
pyinstaller
cbor2
python-decouple